refactor(app-frame): split card_tables.py into four topic mixins (#800)
Decompose the 455-LOC CardTablesHandler grab-bag into four cooperating
mixins over AppFrameProto, mirroring the existing sideboard_guide split.
Pure behaviour-preserving method-grouping move: no logic changed, all 29
methods preserved verbatim, shared state still lives in frame/__init__.py.

- zone_editing.py        ZoneEditingHandlers (zone mutation + transfer)
- card_shortcuts.py      CardShortcutHandlers (keyboard shortcuts)
- card_selection.py      CardSelectionHandlers (selection/focus/routing)
- card_inspector_preview.py CardInspectorPreviewHandlers (preview + #792 printing)

Updated handlers/__init__.py exports and the AppFrame MRO in
frame/__init__.py (four mixins in the slot CardTablesHandler occupied,
still above the panel builders and wx.Frame).
