+++ widgets/frames/app_frame/handlers/__init__.py
@@ -1,7 +1,11 @@
 """Handler mixins composed into :class:`AppFrame`."""
 
 from widgets.frames.app_frame.handlers.app_frame import AppFrameHandlersMixin
-from widgets.frames.app_frame.handlers.card_tables import CardTablesHandler
+from widgets.frames.app_frame.handlers.card_inspector_preview import (
+    CardInspectorPreviewHandlers,
+)
+from widgets.frames.app_frame.handlers.card_selection import CardSelectionHandlers
+from widgets.frames.app_frame.handlers.card_shortcuts import CardShortcutHandlers
 from widgets.frames.app_frame.handlers.child_windows import ChildWindowHandlers
 from widgets.frames.app_frame.handlers.data_loading import DataLoadingHandlers
 from widgets.frames.app_frame.handlers.deck_content import DeckContentHandlers
@@ -18,10 +22,13 @@ from widgets.frames.app_frame.handlers.sideboard_guide_persistence import (
 from widgets.frames.app_frame.handlers.sideboard_guide_record import (
     SideboardGuideRecordHandlers,
 )
+from widgets.frames.app_frame.handlers.zone_editing import ZoneEditingHandlers
 
 __all__ = [
     "AppFrameHandlersMixin",
-    "CardTablesHandler",
+    "CardInspectorPreviewHandlers",
+    "CardSelectionHandlers",
+    "CardShortcutHandlers",
     "ChildWindowHandlers",
     "DataLoadingHandlers",
     "DeckContentHandlers",
@@ -30,4 +37,5 @@ __all__ = [
     "SideboardGuideImportExportHandlers",
     "SideboardGuidePersistenceHandlers",
     "SideboardGuideRecordHandlers",
+    "ZoneEditingHandlers",
 ]
